Refactor story parser
